eval: Split out typval_T dumping functions to nvim/encode.c
